+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Zucchini UI Changelog
 Version 0.0.2-SNAPSHOT
 ----------------------
 
-* Import only passed scenarii (contribution by @pierrick-boule)
+* Search in scenario content
+* Import only passed scenarios (contribution by @pierrick-boule)
 * Use Yarn instead of NPM for Javascript packages management